refactor(cockpit): split capability, content, and runtime helpers
